move to test/library/encryption
